refactor: phase 4 splits + auth middleware + logger (R3, R4, U4a, S1)
Refactors / file splits:
- #R3 src/db/database.ts split into per-domain modules (schema, projects,
      sessions, variables, managed-files, oauth-tokens, oauth-flow-state,
      git-integrations, tool-init-state, sub-agents, mcp-subprocesses);
      CapaDatabase class is now a thin facade delegating to repos
- #R4 src/shared/cache.ts → src/shared/cache/ folder (mirror, snapshot, paths,
      validate, stats)
- #R4 src/shared/plugin-manifest.ts → src/shared/plugin-manifest/ folder
      (detect, cursor-parser, claude-parser, mcp-parser, types-helpers)
- src/shared/tty.ts: shared isColorEnabled() helper used by both cli/ui and
      shared/logger; decouples logger from cli/ui

Security:
- #S1 src/server/auth-middleware.ts: Bearer-token auth on /api/* and MCP route
      when host is not loopback; bypass on 127.0.0.1/::1/localhost; token from
      CAPA_AUTH_TOKEN env or generated at ~/.capa/auth.token; constant-time
      compare; X-Capa-Auth-Token fallback header; querystring tokens rejected

UX:
- #U4 (partial) src/shared/logger.ts uses picocolors instead of hardcoded ANSI;
      honors NO_COLOR/CI/--no-color via shared/tty.ts; direct stdout/stderr
      writes (no console.*); setSink() for testability; CAPA_LOG_LEVEL env wiring
