+++ src/shared/logger.ts
@@ -2,6 +2,10 @@
  * Simple but elegant logger with colors and structured formatting
  */
 
+import { format } from 'util';
+import pc from 'picocolors';
+import { isColorEnabled } from './tty';
+
 export enum LogLevel {
   DEBUG = 0,
   INFO = 1,
@@ -15,9 +19,37 @@ interface LoggerOptions {
   prefix?: string;
 }
 
+interface LogSink {
+  stdout?: NodeJS.WritableStream;
+  stderr?: NodeJS.WritableStream;
+}
+
+function colorize(colorFn: (s: string) => string, value: string): string {
+  return isColorEnabled() ? colorFn(value) : value;
+}
+
+function parseLogLevel(value: string | undefined): LogLevel {
+  switch (value?.toUpperCase()) {
+    case 'DEBUG':
+      return LogLevel.DEBUG;
+    case 'INFO':
+      return LogLevel.INFO;
+    case 'WARN':
+      return LogLevel.WARN;
+    case 'ERROR':
+      return LogLevel.ERROR;
+    case 'SILENT':
+      return LogLevel.SILENT;
+    default:
+      return LogLevel.INFO;
+  }
+}
+
 class Logger {
   private level: LogLevel;
   private prefix: string;
+  private stdout: NodeJS.WritableStream = process.stdout;
+  private stderr: NodeJS.WritableStream = process.stderr;
 
   constructor(options: LoggerOptions = {}) {
     this.level = options.level ?? LogLevel.INFO;
@@ -40,6 +72,14 @@ class Logger {
   }
 
   /**
+   * Override stdout/stderr sinks (for tests or custom output)
+   */
+  setSink(sink: LogSink): void {
+    if (sink.stdout !== undefined) this.stdout = sink.stdout;
+    if (sink.stderr !== undefined) this.stderr = sink.stderr;
+  }
+
+  /**
    * Format timestamp
    */
   private timestamp(): string {
@@ -51,40 +91,49 @@ class Logger {
     return `${hours}:${minutes}:${seconds}.${ms}`;
   }
 
+  private formatLine(message: string, ...args: unknown[]): string {
+    return args.length > 0 ? format(message, ...args) : message;
+  }
+
+  private writeOut(stream: NodeJS.WritableStream, message: string, ...args: unknown[]): void {
+    stream.write(this.formatLine(message, ...args) + '\n');
+  }
+
   /**
    * Format prefix with colors
    */
-  private formatPrefix(levelName: string, color: string): string {
-    const ts = this.timestamp();
+  private formatPrefix(levelName: string, colorFn: (s: string) => string): string {
+    const ts = colorize(pc.gray, this.timestamp());
     const prefix = this.prefix ? ` [${this.prefix}]` : '';
-    return `\x1b[90m${ts}\x1b[0m ${color}${levelName}\x1b[0m${prefix}`;
+    const level = colorize(colorFn, levelName);
+    return `${ts} ${level}${prefix}`;
   }
 
   debug(message: string, ...args: any[]): void {
     if (this.level <= LogLevel.DEBUG) {
-      const prefix = this.formatPrefix('DEBUG', '\x1b[36m'); // Cyan
-      console.log(`${prefix} ${message}`, ...args);
+      const prefix = this.formatPrefix('DEBUG', pc.cyan);
+      this.writeOut(this.stdout, `${prefix} ${message}`, ...args);
     }
   }
 
   info(message: string, ...args: any[]): void {
     if (this.level <= LogLevel.INFO) {
-      const prefix = this.formatPrefix('INFO ', '\x1b[32m'); // Green
-      console.log(`${prefix} ${message}`, ...args);
+      const prefix = this.formatPrefix('INFO ', pc.green);
+      this.writeOut(this.stdout, `${prefix} ${message}`, ...args);
     }
   }
 
   warn(message: string, ...args: any[]): void {
     if (this.level <= LogLevel.WARN) {
-      const prefix = this.formatPrefix('WARN ', '\x1b[33m'); // Yellow
-      console.warn(`${prefix} ${message}`, ...args);
+      const prefix = this.formatPrefix('WARN ', pc.yellow);
+      this.writeOut(this.stderr, `${prefix} ${message}`, ...args);
     }
   }
 
   error(message: string, ...args: any[]): void {
     if (this.level <= LogLevel.ERROR) {
-      const prefix = this.formatPrefix('ERROR', '\x1b[31m'); // Red
-      console.error(`${prefix} ${message}`, ...args);
+      const prefix = this.formatPrefix('ERROR', pc.red);
+      this.writeOut(this.stderr, `${prefix} ${message}`, ...args);
     }
   }
 
@@ -93,15 +142,17 @@ class Logger {
    */
   success(message: string, ...args: any[]): void {
     if (this.level <= LogLevel.INFO) {
-      const prefix = this.formatPrefix('INFO ', '\x1b[32m');
-      console.log(`${prefix} \x1b[32m✓\x1b[0m ${message}`, ...args);
+      const prefix = this.formatPrefix('INFO ', pc.green);
+      const mark = colorize(pc.green, '✓');
+      this.writeOut(this.stdout, `${prefix} ${mark} ${message}`, ...args);
     }
   }
 
   failure(message: string, ...args: any[]): void {
     if (this.level <= LogLevel.ERROR) {
-      const prefix = this.formatPrefix('ERROR', '\x1b[31m');
-      console.error(`${prefix} \x1b[31m✗\x1b[0m ${message}`, ...args);
+      const prefix = this.formatPrefix('ERROR', pc.red);
+      const mark = colorize(pc.red, '✗');
+      this.writeOut(this.stderr, `${prefix} ${mark} ${message}`, ...args);
     }
   }
 
@@ -110,10 +161,13 @@ class Logger {
    */
   http(method: string, path: string, status?: number): void {
     if (this.level <= LogLevel.INFO) {
-      const ts = this.timestamp();
+      const ts = colorize(pc.gray, this.timestamp());
+      const label = colorize(pc.magenta, 'HTTP ');
       const statusStr = status ? ` ${status}` : '';
-      const statusColor = status && status >= 400 ? '\x1b[31m' : '\x1b[32m';
-      console.log(`\x1b[90m${ts}\x1b[0m \x1b[35mHTTP \x1b[0m ${method} ${path}${statusColor}${statusStr}\x1b[0m`);
+      const statusPart = status
+        ? colorize(status >= 400 ? pc.red : pc.green, statusStr)
+        : statusStr;
+      this.stdout.write(`${ts} ${label} ${method} ${path}${statusPart}\n`);
     }
   }
 
@@ -122,10 +176,11 @@ class Logger {
    */
   raw(message: string): void {
     if (this.level <= LogLevel.INFO) {
-      console.log(message);
+      this.stdout.write(message + '\n');
     }
   }
 }
 
 // Default logger instance
 export const logger = new Logger();
+logger.setLevel(parseLogLevel(process.env.CAPA_LOG_LEVEL));
