+++ src/shared/plugin-manifest/index.ts
@@ -1,0 +1,8 @@
+export {
+  detectAndParseManifest,
+  discoverPluginEntries,
+  findPluginInDirectory,
+  type DiscoveredPluginEntry,
+} from './detect';
+
+export { resolvePluginServerDef } from './mcp-parser';
